posix: add test that uses MaxUint64 as the length
  - this forced updating many functions to use uint64 to eliminate much of type casting

+++ cmd/lhsm-plugin-posix/posix/copyat.go
@@ -7,18 +7,18 @@ import "io"
 // Like ReadAt, CopyAt returns non-nil error when writen < n, and at end of file
 // that error is io.EOF.
 // Implementation inspired by io.Copy()
-func CopyAt(dst io.WriterAt, src io.ReaderAt, offset int64, n int64) (written int64, err error) {
-	var bufSize int64 = 1024 * 1024
+func CopyAt(dst io.WriterAt, src io.ReaderAt, offset uint64, n uint64) (written uint64, err error) {
+	var bufSize uint64 = 1024 * 1024
 	if bufSize > n {
 		bufSize = n
 	}
 	buf := make([]byte, bufSize)
 	for written < n {
-		nr, er := src.ReadAt(buf, offset+written)
+		nr, er := src.ReadAt(buf, int64(offset+written))
 		if nr > 0 {
-			nw, ew := dst.WriteAt(buf[0:nr], offset+written)
+			nw, ew := dst.WriteAt(buf[0:nr], int64(offset+written))
 			if nw > 0 {
-				written += int64(nw)
+				written += uint64(nw)
 			}
 			if ew != nil {
 				err = ew
